Fix p4testgen drop-by-default, extract shared extern handler, promote p4testgen to CI
Three improvements:

1. Fix p4testgen PNA model: the p4c fork now initializes DROP=true
   (matching PNA spec semantics). All 33 p4testgen-generated tests
   pass — promoted from manual to CI suite. See p4lang/p4c#5569.

2. Extract shared extern method handler: Register, Hash, Counter,
   Meter, InternetChecksum, Digest, and Random handlers are now in
   handleCommonExternMethod() in ArchitectureHelpers.kt. Eliminates
   the last ~70 lines of PSA/PNA duplication.

3. p4testgen suite runs in CI: 4 PNA programs × symbolic execution
   via Z3, generating ~33 test cases with exhaustive path coverage.
